requirements.txt
Django==4.2
mysqlclient==2.1.1
python-dotenv>=1.0.0
